[C] (+) fix issue when swipped game does not have a cover img (steam fault)
